Ux epic (#7)
Hide sidebar navigation if user hasn't signed in and center the sign-in button vertically
Replace the "menu" sidebar tab label with a user icon
The default route for the app is the new tab page
Move bookmarks from the sidebar into the new tab page, below the text field.
Move the "My publishers" list into the user section of the sidebar. Move "Create Publisher" button to a "+" icon beside the title. Update the "Publishers" tab to "Subscriptions"
Move tabs into a new sidebar tab called "Pages". Each navigation is a new item in the list. No back and forward. New tab is a "+" button next to the "Pages" title that routes to new page.
Move the article menu into the pages list item.