Release bugs (#10)
Packages were built with a hard-coded path to node_modules on the build machine because of fs.readFileSync and import.meta.url in tiny-secp256k1. I replaced it with another secp256k1 package to fix.

I realized that the eslint config file was missing, so I restored it and fixed some linting errors.

Sidebar was not changing to the Pages tab when a new page opened. Fixed.

On Mac, the activate event (in main/index.ts) was firing before the app was ready. I added code to ignore the event if the app isn't yet ready.